add pyproject.toml

+++ ani2xcur/cli/image_magick.py
@@ -11,6 +11,7 @@ from ani2xcur.config import (
 )
 from ani2xcur.logger import get_logger
 from ani2xcur.image_magick_manager import (install_image_magick_windows, install_image_magick_linux, uninstall_image_magick_windows, uninstall_image_magick_linux)
+from ani2xcur.utils import (is_admin_on_windows, is_root_on_linux)
 
 logger = get_logger(
     name=LOGGER_NAME,
@@ -21,10 +22,16 @@ logger = get_logger(
 def install_image_magick(install_path: Annotated[Path | None, typer.Option(help="自定义安装 ImageMagick 的目录", resolve_path=True)] = None) -> None:
     """安装 ImageMagick 到系统中"""
     if sys.platform == "win32":
+        if not is_admin_on_windows():
+            logger.error("当前未使用管理员权限运行 Ani2xcur, 无法安装 ImageMagick, 请使用管理员权限启动 Ani2xcur")
+            sys.exit(1)
         install_image_magick_windows(
             install_path=install_path
         )
     elif sys.platform == "linux":
+        if not is_root_on_linux():
+            logger.error("当前未使用 root 权限运行 Ani2xcur, 无法安装 ImageMagick, 请使用 root 权限启动 Ani2xcur")
+            sys.exit(1)
         install_image_magick_linux()
     else:
         logger.error("不支持的系统: %s", sys.platform)
@@ -33,8 +40,14 @@ def install_image_magick(install_path: Annotated[Path | None, typer.Option(help=
 def uninstall_image_magick() -> None:
     """将 ImageMagick 从系统中卸载"""
     if sys.platform == "win32":
+        if not is_admin_on_windows():
+            logger.error("当前未使用管理员权限运行 Ani2xcur, 无法卸载 ImageMagick, 请使用管理员权限启动 Ani2xcur")
+            sys.exit(1)
         uninstall_image_magick_windows()
     elif sys.platform == "linux":
+        if not is_root_on_linux():
+            logger.error("当前未使用 root 权限运行 Ani2xcur, 无法卸载 ImageMagick, 请使用 root 权限启动 Ani2xcur")
+            sys.exit(1)
         uninstall_image_magick_linux()
     else:
         logger.error("不支持的系统: %s", sys.platform)
